+++ application/controllers/fleet/Welcome.php
@@ -1,0 +1,34 @@
+<?php
+
+defined('BASEPATH') OR exit('No direct script access allowed');
+
+class Welcome extends Application
+{
+
+    /**
+     * Fleet Page for this controller.
+     *
+     * Maps to the following URL
+     * 		http://example.com/
+     * 	- or -
+     * 		http://example.com/welcome/index
+     *
+     * So any other public methods not prefixed with an underscore will
+     * map to /welcome/<method_name>
+     * @see https://codeigniter.com/user_guide/general/urls.html
+     */
+    public function index()
+    {
+
+        $this->load->model('airplanes');
+        $rows = array();
+        $params = array('airplanes'=> $rows);//array of arrays
+        $this->data['airplanes'] = $this->airplanes->all();
+        $this->data['pagetitle'] = 'Fleet';
+        $this->data['pagebody'] = 'fleet';
+        $this->data['pagefooter'] = 'layout/footer';
+        $this->data['pageheader'] = 'layout/header';
+        $this->render();
+    }
+
+}