+++ changelog.md
@@ -22,3 +22,9 @@ Changelog format: Markdown
 
 ## Clint Bock - 2017-10-06 9:36 AM
 - Added the Flights Model and mock data 
+
+## Hansol Lee - 2017-10-06 12:31 PM
+- Added controller for fleet
+- Added controller for fleetdetail
+- Added fleet view page
+- Added fleet detail page